Fixed hash files

+++ include/f/hash.h
@@ -26,9 +26,9 @@ typedef struct hash *hash_t;
 
 hash_t hash_new();
 
-void *hash_set(hash_t h, void *k, size_t ksize, void *v);
-void *hash_get(hash_t h, void *k, size_t ksize);
-void *hash_delete(hash_t h, void *k, size_t ksize);
+void *hash_set(hash_t h, void *k, unsigned int ksize, void *v);
+void *hash_get(hash_t h, void *k, unsigned int ksize);
+void *hash_delete(hash_t h, void *k, unsigned int ksize);
 
 unsigned int hash_size(hash_t h);
